chore(testing): vitest workspace
