+++ package.json
@@ -21,11 +21,11 @@
     "lint:check": "eslint . --cache",
     "lint:fix": "eslint . --cache --fix",
     "svelte:check": "svelte-check",
-    "test:podlet-js": "vitest run -r packages/podlet-js --passWithNoTests --coverage",
-    "test:backend": "vitest run -r packages/backend --passWithNoTests --coverage",
-    "test:frontend": "vitest -c packages/frontend/vite.config.js run packages/frontend --passWithNoTests --coverage",
-    "test:shared": "vitest run -r packages/shared --passWithNoTests --coverage",
-    "test:unit": "npm run test:podlet-js && npm run test:backend && npm run test:shared && npm run test:frontend",
+    "test:podlet-js": "vitest run --project podlet-js",
+    "test:backend":   "vitest run --project quadlet",
+    "test:frontend":  "vitest run --project frontend",
+    "test:shared":    "vitest run --project shared",
+    "test:unit": "vitest run",
     "test:e2e": "cd tests/playwright && npm run test:e2e",
     "test:e2e:smoke": "cd tests/playwright && npm run test:e2e:smoke",
     "typecheck:shared": "tsc --noEmit --project packages/shared",
