+++ packages/frontend/vite.config.ts
@@ -1,0 +1,57 @@
+/**********************************************************************
+ * Copyright (C) 2025 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ***********************************************************************/
+import { join } from 'node:path';
+import * as path from 'node:path';
+import { svelte } from '@sveltejs/vite-plugin-svelte';
+import { defineConfig } from 'vite';
+import { fileURLToPath } from 'node:url';
+import tailwindcss from '@tailwindcss/vite';
+
+const filename = fileURLToPath(import.meta.url);
+const PACKAGE_ROOT = path.dirname(filename);
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  mode: process.env.MODE,
+  root: PACKAGE_ROOT,
+  resolve: {
+    alias: {
+      '/@/': join(PACKAGE_ROOT, 'src') + '/',
+      '/@store/': join(PACKAGE_ROOT, 'src', 'stores') + '/',
+      '/@shared/': join(PACKAGE_ROOT, '../shared') + '/',
+    },
+  },
+  plugins: [tailwindcss(), svelte({ hot: !process.env.VITEST })],
+  optimizeDeps: {
+    exclude: [],
+  },
+  base: '',
+  server: {
+    fs: {
+      strict: true,
+    },
+  },
+  build: {
+    sourcemap: true,
+    outDir: '../backend/media',
+    assetsDir: '.',
+
+    emptyOutDir: true,
+    reportCompressedSize: false,
+  },
+});
