+++ packages/shared/tsconfig.json
@@ -10,5 +10,6 @@
     "allowJs": true,
     "checkJs": true
   },
-  "include": ["src/**/*.d.ts", "src/**/*.ts", "src/**/*.js", "types/*.d.ts", "../../types/**/*.d.ts"]
+  "include": ["src/**/*.d.ts", "src/**/*.ts", "src/**/*.js", "types/*.d.ts", "../../types/**/*.d.ts"],
+  "exclude": ["vite.config.ts", "vitest.config.ts"]
 }
