--- packages/shared/vite.config.js
@@ -1,40 +1,0 @@
-import { join } from 'path';
-import { builtinModules } from 'module';
-
-const PACKAGE_ROOT = __dirname;
-
-/**
- * @type {import('vite').UserConfig}
- * @see https://vitejs.dev/config/
- */
-const config = {
-  mode: process.env.MODE,
-  root: PACKAGE_ROOT,
-  envDir: process.cwd(),
-  resolve: {
-    alias: {
-      '/@/': join(PACKAGE_ROOT, 'src') + '/',
-    },
-  },
-  build: {
-    sourcemap: 'inline',
-    target: 'esnext',
-    outDir: 'dist',
-    assetsDir: '.',
-    minify: process.env.MODE === 'production' ? 'esbuild' : false,
-    lib: {
-      entry: 'src/extension.ts',
-      formats: ['cjs'],
-    },
-    rollupOptions: {
-      external: ['@podman-desktop/api', ...builtinModules.flatMap(p => [p, `node:${p}`])],
-      output: {
-        entryFileNames: '[name].js',
-      },
-    },
-    emptyOutDir: true,
-    reportCompressedSize: false,
-  },
-};
-
-export default config;
