+++ vitest.workspace.ts
@@ -15,25 +15,8 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
-import path from 'node:path';
-import { join } from 'path';
+import { defineWorkspace } from 'vitest/config';
 
-const PACKAGE_ROOT = __dirname;
-
-const config = {
-  test: {
-    include: ['**/*.{test,spec}.?(c|m)[jt]s?(x)'],
-    coverage: {
-      provider: 'v8',
-      reporter: ['lcov', 'text'],
-    },
-  },
-  resolve: {
-    alias: {
-      '@podman-desktop/api': path.resolve(__dirname, '__mocks__/@podman-desktop/api.js'),
-      '/@/': join(PACKAGE_ROOT, 'src') + '/',
-    },
-  },
-};
-
-export default config;
+export default defineWorkspace([
+  'packages/**/vitest.config.{js,ts}',
+]);
